+++ daemon.php
@@ -8,6 +8,7 @@ use Movim\Bootstrap;
 use Movim\Console\DaemonCommand;
 use Movim\Console\ConfigCommand;
 use Movim\Console\EmojisToJsonCommand;
+use Movim\Console\CompileLanguages;
 use Symfony\Component\Console\Application;
 
 $bootstrap = new Bootstrap;
@@ -17,4 +18,5 @@ $application = new Application;
 $application->add(new ConfigCommand);
 $application->add(new DaemonCommand);
 $application->add(new EmojisToJsonCommand);
+$application->add(new CompileLanguages);
 $application->run();