+++ src/Movim/Console/ConfigCommand.php
@@ -99,6 +99,6 @@ class ConfigCommand extends Command
             }
         }
 
-        return 1;
+        return 0;
     }
 }